requirements.txt
streamlit
pandas
numpy
PyYAML
yfinance
matplotlib
reportlab
openpyxl